Refactoring and reassembly
Work has been done to rework the created code, simplifying it in the name of simplicity. Previously missing methods have been added, each function has a footnote in the code with the name from the API in Russian. There are plans to add documentation...

Each function retains the original values given in the documentation on the Digiseller website

The id_i parameters of the message-related functions have been renamed to order_id